feat: Implement Redis caching optimization for model responses

+++ weaver_ai/models/mock.py
@@ -47,7 +47,7 @@ class MockAdapter:
             elif op == "*":
                 result = a * b
             elif op == "/" and b != 0:
-                result = a / b
+                result = int(a / b)
             else:
                 return "Cannot evaluate this expression"
